pyproject.toml
[project]
name = "voicebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "edge-tts>=7.2.3",
    "fastapi>=0.123.7",
    "google-genai>=1.52.0",
    "gtts",
    "numpy>=2.3.5",
    "pinecone>=8.0.0",
    "playsound",
    "pydub>=0.25.1",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "pyttsx3>=2.99",
    "rank-bm25>=0.2.2",
    "rich>=14.2.0",
    "scipy>=1.16.3",
    "sentence-transformers>=5.1.2",
    "sentencepiece>=0.2.1",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
    "speechrecognition>=3.14.4",
    "uvicorn>=0.38.0",
]
